feat(dto): change dtos namespace
dto id moved to abstarctions
Update mapping references to align with the new DTO structure.
