+++ ServiceStation/DataTransferObjects/Abstraction/AbstractDto.cs
@@ -1,0 +1,6 @@
+﻿namespace ServiceStation.DataTransferObjects.Abstraction;
+
+public abstract class AbstractDto
+{
+    public Guid Id { get; init; }
+}
