--- ServiceStation/Models/DTOs/Abstraction/AbstractDto.cs
@@ -1,5 +1,0 @@
-﻿namespace ServiceStation.Models.DTOs.Abstraction;
-
-public abstract class AbstractDto
-{
-}
