+++ ServiceStation/Services/Mapping/Abstraction/IMapper.cs
@@ -1,4 +1,4 @@
-﻿using ServiceStation.Models.DTOs.Abstraction;
+﻿using ServiceStation.DataTransferObjects.Abstraction;
 using ServiceStation.Models.Entities.Abstraction;
 
 namespace ServiceStation.Services.Mapping.Abstraction;
