+++ ServiceStation/Services/Mapping/Implementation/DefectMapper.cs
@@ -1,4 +1,4 @@
-﻿using ServiceStation.Models.DTOs.Implementation;
+﻿using ServiceStation.DataTransferObjects.Implementation;
 using ServiceStation.Models.Entities.Implementation;
 using ServiceStation.Services.Mapping.Abstraction;
