+++ ServiceStation/Services/Mapping/Implementation/MapperExtensions.cs
@@ -1,4 +1,4 @@
-﻿using ServiceStation.Models.DTOs.Abstraction;
+﻿using ServiceStation.DataTransferObjects.Abstraction;
 using ServiceStation.Models.Entities.Abstraction;
 using ServiceStation.Services.Mapping.Abstraction;
