+++ ServiceStation/Services/Mapping/Implementation/OwnerMapper.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text;
-using ServiceStation.Models.DTOs.Implementation;
+using ServiceStation.DataTransferObjects.Implementation;
 using ServiceStation.Models.Entities.Implementation;
 using ServiceStation.Services.Mapping.Abstraction;
