+++ ServiceStation/ViewModels/Implementation/OwnerDetailsViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text;
 using Microsoft.Extensions.Logging;
-using ServiceStation.Models.DTOs.Implementation;
+using ServiceStation.DataTransferObjects.Implementation;
 using ServiceStation.Models.Entities.Implementation;
 using ServiceStation.Repository.Abstraction;
 using ServiceStation.Services.Mapping.Abstraction;
